requirements.txt
asgiref==3.10.0
astroid==4.0.2
colorama==0.4.6
coverage==7.12.0
dill==0.4.0
Django==5.2.7
et_xmlfile==2.0.0
flake8==7.3.0
fpdf==1.7.2
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.3
pytest==9.0.1
pytest-cov==7.0.0
sqlparse==0.5.3
tomlkit==0.13.3
tzdata==2025.2
